fix(ai): remove vitest dependency from test exports
