+++ .changeset/chatty-rocks-rest.md
@@ -1,0 +1,5 @@
+---
+'ai': patch
+---
+
+fix(ai): remove vitest dependency from test exports
