+++ packages/ai/src/test/mock-embedding-model-v2.ts
@@ -1,8 +1,5 @@
 import { EmbeddingModelV2 } from '@ai-sdk/provider';
-import { Embedding } from '../types';
-import { EmbeddingModelUsage } from '../types/usage';
 import { notImplemented } from './not-implemented';
-import { assert } from 'vitest';
 
 export class MockEmbeddingModelV2<VALUE> implements EmbeddingModelV2<VALUE> {
   readonly specificationVersion = 'v2';
@@ -35,21 +32,4 @@ export class MockEmbeddingModelV2<VALUE> implements EmbeddingModelV2<VALUE> {
     this.supportsParallelCalls = supportsParallelCalls;
     this.doEmbed = doEmbed;
   }
-}
-
-export function mockEmbed<VALUE>(
-  expectedValues: Array<VALUE>,
-  embeddings: Array<Embedding>,
-  usage?: EmbeddingModelUsage,
-  response: Awaited<
-    ReturnType<EmbeddingModelV2<VALUE>['doEmbed']>
-  >['response'] = { headers: {}, body: {} },
-  providerMetadata?: Awaited<
-    ReturnType<EmbeddingModelV2<VALUE>['doEmbed']>
-  >['providerMetadata'],
-): EmbeddingModelV2<VALUE>['doEmbed'] {
-  return async ({ values }) => {
-    assert.deepStrictEqual(expectedValues, values);
-    return { embeddings, usage, response, providerMetadata };
-  };
 }
